FRACTRAN

Documentation for FRACTRAN.

FRACTRAN._factorizations

FRACTRAN.primes

FRACTRAN.add

FRACTRAN.div

FRACTRAN.factorize

FRACTRAN.fractran

FRACTRAN.generate_primes

FRACTRAN.mul

FRACTRAN.prettify_factorization

FRACTRAN.primegame

FRACTRAN.sub

FRACTRAN._factorizations — Constant

Non-public cache for yet computed factorizations.

FRACTRAN.primes — Constant

All yet computed prime numbers.

FRACTRAN.add — Method

add(a::Int64, b::Int64)::Int64

Add b to a using the following FRACTRAN program:

Start value: n = 2^a * 3^b

Fractions: 3//2

Result: 3^(a+b)

Note: add(a, b) directly returns a + b, not 3^(a+b).

FRACTRAN.div — Method

div(a::Int64, b::Int64)::Rational{Int64}

Divide a by b using the following FRACTRAN program:

Start value: 2^a * 3^b * 11

Fractions: 91//66, 11//13, 1//33, 85//11, 57//119, 17//19, 11//17, 1//3

Result: 5^q * 7^r (q: quotient, r: remainder)

Note: div(a, b) directly returns a // b, not 5^q * 7^r.

FRACTRAN.factorize — Method

factorize(n::Int64)::Accumulator{Int64, Int64}

Factorize n. This yields an Accumulator mapping the factors' bases to their counts (exponents).

FRACTRAN.fractran — Method

fractran(
    n::Int64,
    fractions::Rational{Int64}...;
    return_first::Bool = false,
)::Int64

fractran(
    n::Int64,
    fractions::Tuple{Vararg{Rational{Int64}}};
    return_first::Bool = false,
)::Int64

Run the FRACTRAN algorithm on the start value n using the fractions. Iff return_first is true, return the first obtained integer. Else, run the algorithm until no fraction yields an integer and return the last integer. This is the FRACTRAN algorithm's actual result—return_first is needed by some specific programs like PRIMEGAME that filter the FRACTRAN integers.

FRACTRAN.generate_primes — Method

generate_primes(min_n::Int64, max_n::Int64)::Vector{Int64}
generate_primes(max_n::Int64)::Vector{Int64}

Compute the prime numbers from min_n to max_n, inclusive. The second form defaults to min_n = 2.

FRACTRAN.mul — Method

mul(a::Int64, b::Int64)::Int64

Multiply a by b using the following FRACTRAN program:

Start value: n = 2^a * 3^b

Fractions: 455//33, 11//13, 1//11, 3//7, 11//2, 1//3

Result: 5^(a*b)

Note: mul(a, b) directly returns a * b, not 5^(a*b).

FRACTRAN.prettify_factorization — Method

prettify_factorization(factors::Accumulator{Int64, Int64})::String

Stringify the factors' factorization in the canonical, condensed way.

FRACTRAN.primegame — Function

primegame(max_iterations::Int64 = 100)::Vector{Int64}

Find all prime numbers that are reachable by running max_iterations iterations of the algorithm using the following FRACTRAN program, called "PRIMEGAME":

Start value: n = 2 (in PRIMEGAME, more generally, n = 2^a * 7^b)

Fractions: 17//91, 78//85, 19//51, 23//38, 29//33, 77//29, 95//23, 77//19, 1//17, 11//13, 13//11, 15//2, 1//7, 55//1

Result: 2^c * 7^d, with c ≤ a and d ≤ b, a prime number iff d == 0

Note: primegame(n) directly returns all prime numbers up to, and including, c, not 2^c * 7^d. Also note that the PRIMEGAME algorithm is very inefficient and may take a very long time even for small prime numbers.

FRACTRAN.sub — Method

sub(a::Int64, b::Int64)::Int64

Subtract b from a using the following FRACTRAN program:

Start value: n = 2^a * 3^b

Fractions: 1//6

Result: 2^(a-b)

Note: sub(a, b) directly returns a - b, not 2^(a-b).