API reference

FRACTRAN.jl offers the below functions and cache variables within the FRACTRAN module namespace as public interface. All undocumented symbols are considered private implementation details and may change at any time. Currently, there are very few of which, but should there be one that you'd like to become public, please open an issue asking for it.

Note

The docstrings for the non-mutating and mutating versions of a function are identical, but included in duplicated form for searchability. So don't read both and try to find a difference…

Table of contents

Table of contents

Index

Module

Prime factorization

FRACTRAN algorithm

Example programs

Addition program

Subtraction program

Multiplication program

Division program

PRIMEGAME program

Cache variables

Index

The following symbols in FRACTRAN.jl are exported or public:

FRACTRAN.FRACTRAN

FRACTRAN.add

FRACTRAN.add!

FRACTRAN.divrem

FRACTRAN.divrem!

FRACTRAN.factorize

FRACTRAN.factorize!

FRACTRAN.fractran

FRACTRAN.fractran!

FRACTRAN.generate_primes

FRACTRAN.mul

FRACTRAN.mul!

FRACTRAN.prettify_factorization

FRACTRAN.primegame

FRACTRAN.primegame!

FRACTRAN.sub

FRACTRAN.sub!

FRACTRAN.factorizations

FRACTRAN.primes

Module

FRACTRAN.FRACTRAN — Module

Julia implementation of the esoteric programming language FRACTRAN.

Extended help

Algorithm

FRACTRAN is an esoteric programming language developed by John Conway in 1987. The algorithm behind FRACTRAN is based on multiplying a natural number (the start number $n$) successively by fractions ($f$), until a new natural number is obtained, the algorithm's result. FRACTRAN.jl implements this algorithm in the fractran function.

For a more extensive overview over FRACTRAN, refer to the FRACTRAN.jl documentation, the Wikipedia entry, or the Esolang wiki.

Note

FRACTRAN is rather slow, especially more complex programs like PRIMEGAME.

Warning

Rather do not use FRACTRAN in performance-critical applications or production environments, just for educational purposes.

Tip

Unless you need a break…

Module contents

FRACTRAN.jl provides both an implementation of the algorithm itself, as well as several example FRACTRAN programs. Additionally, the module contains a prime number generator and a factorization function for factorizing the numbers for FRACTRAN.

The actual FRACTRAN implementation lies within the fractran function. For using FRACTRAN, you only need to pass the start number and fractions to fractran, and get the algorithm's result back as integer.

To show the possibilities of FRACTRAN (it's Turing-complete, after all), FRACTRAN.jl provides some (public, not exported) example FRACTRAN functions: FRACTRAN.add, FRACTRAN.sub, FRACTRAN.mul, and FRACTRAN.divrem for the respective arithmetic operations. A fifth function, FRACTRAN.primegame, implements the famous PRIMEGAME program, which generates prime numbers. You may also want to have a look at the source code to see how to programmatically use FRACTRAN in Julia.

Finally, as FRACTRAN is prime number–based, FRACTRAN.jl also includes a function for prime factorization, factorize, which uses generate_primes to generate a list of prime numbers.

Note

Though you can use generate_primes and factorize for other purposes, the implementation is (intentionally) not optimized for performance.

Tip

You may visualize the factorization with prettify_factorization.

To accelerate repeated fractran calls, factorize, fractran, and all example programs have a second form with exclamation mark (like factorize!, fractran! etc.) that take cache arguments for in-place mutation. FRACTRAN.jl provides two module-level caches for this, FRACTRAN.factorizations and FRACTRAN.primes, which you can use alongside your own objects.

Examples

Manual addition program

julia> n = 432;                # Start number: 2^4 * 3^3 (operands 4 and 3).julia> fractions = (3//2,);    # FRACTRAN program for addition.julia> fractran(n, fractions)  # FRACTRAN invocation.2187julia> factorize(ans)          # Get exponents, result is in register `3`.DataStructures.Accumulator{Int64, Int64} with 1 entry:  3 => 7

FRACTRAN.jl example programs

julia> FRACTRAN.add(4, 3)7julia> FRACTRAN.sub(4, 3)1julia> FRACTRAN.mul(4, 3)12julia> FRACTRAN.divrem(4, 3)(1, 1)julia> FRACTRAN.primegame(300)  # Number of FRACTRAN iterations.3-element Vector{Int64}: 2 3 5

Prime number generation and factorization

julia> generate_primes(2, 10)4-element Vector{Int64}: 2 3 5 7julia> factorize(120)DataStructures.Accumulator{Int64, Int64} with 3 entries:  3 => 1  2 => 3  5 => 1julia> prettify_factorization(ans)"2³⋅3⋅5"

Prime factorization

To simplify the underlying computations, FRACTRAN.jl implements FRACTRAN using prime factorization by factorize (non-cache-mutating) and factorize! (cache-mutating). Thereby, generate_primes generates the needed list of prime numbers. Additionally, prettify_factorization can create a prettified string representation of a factorization for prettyprinting. All functions are exported for external usage by you.

FRACTRAN.generate_primes — Function

generate_primes(min_n::Integer, max_n::Integer)::Vector{Int}
generate_primes(max_n::Integer)::Vector{Int}

Compute the prime numbers from min_n to max_n, inclusive.

The second form defaults to min_n = 2.

Examples

julia> generate_primes(2, 20)8-element Vector{Int64}:  2  3  5  7 11 13 17 19julia> generate_primes(10)     # Usage of default value 2 for `min_n`.4-element Vector{Int64}: 2 3 5 7julia> generate_primes(-5, 1)  # There's no prime number below 2.ERROR: ArgumentError:  min_n must be ≥ 2.[...]

FRACTRAN.factorize — Function

factorize(n::Integer)::Accumulator{Int, Int}

factorize!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
)::Accumulator{Int, Int}

Factorize n to prime factors.

The factorization yields an Accumulator mapping the factors' bases to their exponents (counts). The second form takes cache arguments for accelerated computations, see the Extended help.

Examples

julia> factorize(120)  # Composite number.DataStructures.Accumulator{Int64, Int64} with 3 entries:  3 => 1  2 => 3  5 => 1julia> factorize(7)    # Prime number.DataStructures.Accumulator{Int64, Int64} with 1 entry:  7 => 1julia> factorize(1)    # Empty product.Accumulator{Int64,Int64}()

Extended help

Unlike the non-mutating factorize, factorize! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call factorize repeatedly, it is thus more efficient to call factorize! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.factorize! — Function

factorize(n::Integer)::Accumulator{Int, Int}

factorize!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
)::Accumulator{Int, Int}

Factorize n to prime factors.

The factorization yields an Accumulator mapping the factors' bases to their exponents (counts). The second form takes cache arguments for accelerated computations, see the Extended help.

Examples

julia> factorize(120)  # Composite number.DataStructures.Accumulator{Int64, Int64} with 3 entries:  3 => 1  2 => 3  5 => 1julia> factorize(7)    # Prime number.DataStructures.Accumulator{Int64, Int64} with 1 entry:  7 => 1julia> factorize(1)    # Empty product.Accumulator{Int64,Int64}()

Extended help

Unlike the non-mutating factorize, factorize! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call factorize repeatedly, it is thus more efficient to call factorize! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.prettify_factorization — Function

prettify_factorization(
    factors::Accumulator{Int, Int};
    explicit_one::Bool = false,
    verbose::Bool = false,
)::String

Stringify the factors' factorization in the canonical, condensed way.

If explicit_one is true (default: false), also print exponents of one (1), as ¹. If verbose is true (default: false), expand the exponents to individual factors. In this case, explicit_one has no effect.

Examples

julia> factors = factorize(120)DataStructures.Accumulator{Int64, Int64} with 3 entries:  3 => 1  2 => 3  5 => 1julia> prettify_factorization(factors)"2³⋅3⋅5"julia> prettify_factorization(factorize(120))  # Same as above."2³⋅3⋅5"julia> prettify_factorization(factorize(120), explicit_one=true)"2³⋅3¹⋅5¹"julia> prettify_factorization(factorize(120), verbose=true)"2⋅2⋅2⋅3⋅5"

FRACTRAN algorithm

The FRACTRAN algorithm is implemented in fractran and its in-place cache-mutating version fractran!. Both functions are exported.

FRACTRAN.fractran — Function

fractran(
    n::Integer,
    fractions::Rational{<:Integer}...;
    return_first::Bool = false,
)::Int

fractran(
    n::Integer,
    fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}};
    return_first::Bool = false,
)::Int

fractran!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
    fractions::Rational{<:Integer}...;
    return_first::Bool = false,
)::Int

fractran!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
    fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}};
    return_first::Bool = false,
)::Int

Run the FRACTRAN algorithm on the start value n using the fractions.

If return_first is true (default: false), return the first obtained integer. Else, run the algorithm until no fraction yields an integer and return the last integer. This is the FRACTRAN algorithm's actual result—return_first is needed by some specific programs like PRIMEGAME that filter the FRACTRAN integers.

The third and fourth forms take cache arguments for accelerated computations, see the Extended help.

Examples

Simple addition program

julia> n = 432;                # Start number: 2^4 * 3^3 (operands 4 and 3).julia> fractions = (3//2,);    # FRACTRAN program for addition.julia> fractran(n, fractions)  # FRACTRAN invocation.2187julia> factorize(ans)          # Get exponents, result is in register `3`.DataStructures.Accumulator{Int64, Int64} with 1 entry:  3 => 7

Subtraction program with cache

julia> using DataStructures: DataStructuresjulia> factorizations = Dict{Int, DataStructures.Accumulator{Int, Int}}();julia> primes = Int[];julia> n = 432;                # Start number: 2^4 * 3^3 (operands 4 and 3).julia> fractions = (1//6,);    # FRACTRAN program for subtraction.julia> fractran(n, fractions)  # FRACTRAN invocation.2julia> factorize(ans)          # Get exponents, result is in register `2`.DataStructures.Accumulator{Int64, Int64} with 1 entry:  2 => 1

Extended help

Unlike the non-mutating fractran, fractran! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call fractran repeatedly, it is thus more efficient to call fractran! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.fractran! — Function

fractran(
    n::Integer,
    fractions::Rational{<:Integer}...;
    return_first::Bool = false,
)::Int

fractran(
    n::Integer,
    fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}};
    return_first::Bool = false,
)::Int

fractran!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
    fractions::Rational{<:Integer}...;
    return_first::Bool = false,
)::Int

fractran!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    n::Integer,
    fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}};
    return_first::Bool = false,
)::Int

Run the FRACTRAN algorithm on the start value n using the fractions.

If return_first is true (default: false), return the first obtained integer. Else, run the algorithm until no fraction yields an integer and return the last integer. This is the FRACTRAN algorithm's actual result—return_first is needed by some specific programs like PRIMEGAME that filter the FRACTRAN integers.

The third and fourth forms take cache arguments for accelerated computations, see the Extended help.

Examples

Simple addition program

julia> n = 432;                # Start number: 2^4 * 3^3 (operands 4 and 3).julia> fractions = (3//2,);    # FRACTRAN program for addition.julia> fractran(n, fractions)  # FRACTRAN invocation.2187julia> factorize(ans)          # Get exponents, result is in register `3`.DataStructures.Accumulator{Int64, Int64} with 1 entry:  3 => 7

Subtraction program with cache

julia> using DataStructures: DataStructuresjulia> factorizations = Dict{Int, DataStructures.Accumulator{Int, Int}}();julia> primes = Int[];julia> n = 432;                # Start number: 2^4 * 3^3 (operands 4 and 3).julia> fractions = (1//6,);    # FRACTRAN program for subtraction.julia> fractran(n, fractions)  # FRACTRAN invocation.2julia> factorize(ans)          # Get exponents, result is in register `2`.DataStructures.Accumulator{Int64, Int64} with 1 entry:  2 => 1

Extended help

Unlike the non-mutating fractran, fractran! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call fractran repeatedly, it is thus more efficient to call fractran! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

Example programs

FRACTRAN.jl bundles several example programs to showcase FRACTRAN. All of them are public to prevent cluttering your namespace when using FRACTRAN. All functions have a second form with exclamation mark that operate in-place on provided cache arguments.

Addition program

FRACTRAN.add — Function

add(a::Integer, b::Integer)::Int

add!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Add b to a using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 3//2

Result: 3^(a+b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

add(a, b) directly returns a + b, not 3^(a+b).

Examples

julia> FRACTRAN.add(4, 3)7julia> FRACTRAN.add(1, 2)3

Extended help

Unlike the non-mutating add, add! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call add repeatedly, it is thus more efficient to call add! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.add! — Function

add(a::Integer, b::Integer)::Int

add!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Add b to a using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 3//2

Result: 3^(a+b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

add(a, b) directly returns a + b, not 3^(a+b).

Examples

julia> FRACTRAN.add(4, 3)7julia> FRACTRAN.add(1, 2)3

Extended help

Unlike the non-mutating add, add! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call add repeatedly, it is thus more efficient to call add! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

Subtraction program

FRACTRAN.sub — Function

sub(a::Integer, b::Integer)::Int

sub!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Subtract b from a using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 1//6

Result: 2^(a-b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

sub(a, b) directly returns a - b, not 2^(a-b).

Examples

julia> FRACTRAN.sub(4, 3)1julia> FRACTRAN.sub(1, 2)-1

Extended help

Unlike the non-mutating sub, sub! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call sub repeatedly, it is thus more efficient to call sub! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.sub! — Function

sub(a::Integer, b::Integer)::Int

sub!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Subtract b from a using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 1//6

Result: 2^(a-b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

sub(a, b) directly returns a - b, not 2^(a-b).

Examples

julia> FRACTRAN.sub(4, 3)1julia> FRACTRAN.sub(1, 2)-1

Extended help

Unlike the non-mutating sub, sub! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call sub repeatedly, it is thus more efficient to call sub! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

Multiplication program

FRACTRAN.mul — Function

mul(a::Integer, b::Integer)::Int

mul!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Multiply a by b using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 455//33, 11//13, 1//11, 3//7, 11//2, 1//3

Result: 5^(a*b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

mul(a, b) directly returns a * b, not 5^(a*b).

Examples

julia> FRACTRAN.mul(4, 3)12julia> FRACTRAN.mul(1, 2)2

Extended help

Unlike the non-mutating mul, mul! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call mul repeatedly, it is thus more efficient to call mul! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.mul! — Function

mul(a::Integer, b::Integer)::Int

mul!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Int

Multiply a by b using a FRACTRAN program.

The program is:

Start value: n = 2^a * 3^b

Fractions: 455//33, 11//13, 1//11, 3//7, 11//2, 1//3

Result: 5^(a*b)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

mul(a, b) directly returns a * b, not 5^(a*b).

Examples

julia> FRACTRAN.mul(4, 3)12julia> FRACTRAN.mul(1, 2)2

Extended help

Unlike the non-mutating mul, mul! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call mul repeatedly, it is thus more efficient to call mul! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

Division program

FRACTRAN.divrem — Function

divrem(a::Integer, b::Integer)::Tuple{Int, Int}

divrem!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Tuple{Int, Int}

Divide a by b (as Euclidian division) using a FRACTRAN program.

The program is:

Start value: 2^a * 3^b * 11

Fractions: 91//66, 11//13, 1//33, 85//11, 57//119, 17//19, 11//17, 1//3

Result: 5^q * 7^r (q ≔ a ÷ b: quotient, r ≔ a % b: remainder)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

divrem(a, b) directly returns (q, r), i.e., (a ÷ b, a % b), not 5^q * 7^r.

Examples

julia> FRACTRAN.divrem(4, 3)(1, 1)julia> FRACTRAN.divrem(1, 2)(0, 1)julia> FRACTRAN.divrem(2, 1)(2, 0)

Extended help

Unlike the non-mutating divrem, divrem! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call divrem repeatedly, it is thus more efficient to call divrem! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.divrem! — Function

divrem(a::Integer, b::Integer)::Tuple{Int, Int}

divrem!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    a::Integer,
    b::Integer,
)::Tuple{Int, Int}

Divide a by b (as Euclidian division) using a FRACTRAN program.

The program is:

Start value: 2^a * 3^b * 11

Fractions: 91//66, 11//13, 1//33, 85//11, 57//119, 17//19, 11//17, 1//3

Result: 5^q * 7^r (q ≔ a ÷ b: quotient, r ≔ a % b: remainder)

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

divrem(a, b) directly returns (q, r), i.e., (a ÷ b, a % b), not 5^q * 7^r.

Examples

julia> FRACTRAN.divrem(4, 3)(1, 1)julia> FRACTRAN.divrem(1, 2)(0, 1)julia> FRACTRAN.divrem(2, 1)(2, 0)

Extended help

Unlike the non-mutating divrem, divrem! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call divrem repeatedly, it is thus more efficient to call divrem! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

PRIMEGAME program

FRACTRAN.primegame — Function

primegame(max_iterations::Integer = 100)::Vector{Int}

primegame!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    max_iterations::Integer = 100,
)::Vector{Int}

Find prime numbers using the FRACTRAN program "PRIMEGAME".

The prime numbers returned are all that are reachable by running max_iterations (default: 100) iterations of the FRACTRAN algorithm, returned in sorted order, starting with 2.

The program is:

Start value: n = 2 (in PRIMEGAME, more generally, n = 2^a * 7^b)

Fractions: 17//91, 78//85, 19//51, 23//38, 29//33, 77//29, 95//23, 77//19, 1//17, 11//13, 13//11, 15//2, 1//7, 55//1

Result: 2^c * 7^d, with c ≤ a and d ≤ b, a prime number iff d == 0

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

primegame(n) directly returns all prime numbers up to, and including, c, not 2^c * 7^d.

Warning

The PRIMEGAME algorithm is very inefficient and may take a very long time even for small prime numbers (e.g. $1000$ iterations find just four prime numbers).

Examples

julia> FRACTRAN.primegame(300)  # Number of FRACTRAN iterations.3-element Vector{Int64}: 2 3 5julia> FRACTRAN.primegame()     # Usage of default value 100.2-element Vector{Int64}: 2 3

Extended help

Unlike the non-mutating primegame, primegame! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call primegame repeatedly, it is thus more efficient to call primegame! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

FRACTRAN.primegame! — Function

primegame(max_iterations::Integer = 100)::Vector{Int}

primegame!(
    factorizations::Dict{Int, Accumulator{Int, Int}},
    primes::Vector{Int},
    max_iterations::Integer = 100,
)::Vector{Int}

Find prime numbers using the FRACTRAN program "PRIMEGAME".

The prime numbers returned are all that are reachable by running max_iterations (default: 100) iterations of the FRACTRAN algorithm, returned in sorted order, starting with 2.

The program is:

Start value: n = 2 (in PRIMEGAME, more generally, n = 2^a * 7^b)

Fractions: 17//91, 78//85, 19//51, 23//38, 29//33, 77//29, 95//23, 77//19, 1//17, 11//13, 13//11, 15//2, 1//7, 55//1

Result: 2^c * 7^d, with c ≤ a and d ≤ b, a prime number iff d == 0

The second form takes cache arguments for accelerated computations, see the Extended help.

Note

primegame(n) directly returns all prime numbers up to, and including, c, not 2^c * 7^d.

Warning

The PRIMEGAME algorithm is very inefficient and may take a very long time even for small prime numbers (e.g. $1000$ iterations find just four prime numbers).

Examples

julia> FRACTRAN.primegame(300)  # Number of FRACTRAN iterations.3-element Vector{Int64}: 2 3 5julia> FRACTRAN.primegame()     # Usage of default value 100.2-element Vector{Int64}: 2 3

Extended help

Unlike the non-mutating primegame, primegame! takes cache arguments of pre-computed factorizations and primes and mutates them in-place, thus updating the cache for future usage. When needing to call primegame repeatedly, it is thus more efficient to call primegame! instead and pass shared factorizations and primes arguments.

To this end, FRACTRAN.jl provides two module-level cache variables, FRACTRAN.factorizations and FRACTRAN.primes, which you can use as storage targets.

Warning

Using the module-level cache variables is not thread-safe.

Cache variables

The two variables FRACTRAN.factorizations and FRACTRAN.primes serve as module-level caches for repeated calls to fractran and are declared public.

FRACTRAN.factorizations — Constant

factorizations::Dict{Int, Accumulator{Int, Int}}

Module-level cache for yet computed factorizations.

Extended help

This optional cache can be useful to accelerate repeated fractran calls when using FRACTRAN.jl's mutating functions.

These are:

factorize!

fractran!

add!

sub!

mul!

divrem!

primegame!

FRACTRAN.primes — Constant

primes::Vector{Int}

Module-level cache for yet computed prime numbers.

Extended help

This optional cache can be useful to accelerate repeated fractran calls when using FRACTRAN.jl's mutating functions.

These are:

factorize!

fractran!

add!

sub!

mul!

divrem!

primegame!

add(a::Integer, b::Integer)::Int add!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Add b to a using a FRACTRAN program.

add(a::Integer, b::Integer)::Int add!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Add b to a using a FRACTRAN program.

divrem(a::Integer, b::Integer)::Tuple{Int, Int} divrem!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Tuple{Int, Int}
Divide a by b (as Euclidian division) using a FRACTRAN program.

divrem(a::Integer, b::Integer)::Tuple{Int, Int} divrem!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Tuple{Int, Int}
Divide a by b (as Euclidian division) using a FRACTRAN program.

factorize(n::Integer)::Accumulator{Int, Int} factorize!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, )::Accumulator{Int, Int}
Factorize n to prime factors.

factorize(n::Integer)::Accumulator{Int, Int} factorize!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, )::Accumulator{Int, Int}
Factorize n to prime factors.

fractran( n::Integer, fractions::Rational{<:Integer}...; return_first::Bool = false, )::Int fractran( n::Integer, fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}}; return_first::Bool = false, )::Int fractran!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, fractions::Rational{<:Integer}...; return_first::Bool = false, )::Int fractran!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}}; return_first::Bool = false, )::Int
Run the FRACTRAN algorithm on the start value n using the fractions.

fractran( n::Integer, fractions::Rational{<:Integer}...; return_first::Bool = false, )::Int fractran( n::Integer, fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}}; return_first::Bool = false, )::Int fractran!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, fractions::Rational{<:Integer}...; return_first::Bool = false, )::Int fractran!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, n::Integer, fractions::Tuple{Rational{<:Integer}, Vararg{Rational{<:Integer}}}; return_first::Bool = false, )::Int
Run the FRACTRAN algorithm on the start value n using the fractions.

generate_primes(min_n::Integer, max_n::Integer)::Vector{Int} generate_primes(max_n::Integer)::Vector{Int}
Compute the prime numbers from min_n to max_n, inclusive.

mul(a::Integer, b::Integer)::Int mul!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Multiply a by b using a FRACTRAN program.

mul(a::Integer, b::Integer)::Int mul!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Multiply a by b using a FRACTRAN program.

prettify_factorization( factors::Accumulator{Int, Int}; explicit_one::Bool = false, verbose::Bool = false, )::String
Stringify the factors' factorization in the canonical, condensed way.

primegame(max_iterations::Integer = 100)::Vector{Int} primegame!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, max_iterations::Integer = 100, )::Vector{Int}
Find prime numbers using the FRACTRAN program "PRIMEGAME".

primegame(max_iterations::Integer = 100)::Vector{Int} primegame!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, max_iterations::Integer = 100, )::Vector{Int}
Find prime numbers using the FRACTRAN program "PRIMEGAME".

sub(a::Integer, b::Integer)::Int sub!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Subtract b from a using a FRACTRAN program.

sub(a::Integer, b::Integer)::Int sub!( factorizations::Dict{Int, Accumulator{Int, Int}}, primes::Vector{Int}, a::Integer, b::Integer, )::Int
Subtract b from a using a FRACTRAN program.